+++ src/Pinventory.Pins.Import.Worker/DebugWolverineExtensions.cs
@@ -2,7 +2,7 @@
 using Pinventory.Pins.Application.Tagging.Messages;
 using Pinventory.Pins.Domain.Importing;
 using Pinventory.Pins.Domain.Importing.Events;
-using Pinventory.Pins.Infrastructure.Sagas.Messages;
+using Pinventory.Pins.Infrastructure.Importing.Sagas.Messages;
 using Pinventory.ServiceDefaults.Wolverine;
 
 namespace Pinventory.Pins.Import.Worker;
