+++ src/Pinventory.Pins.Infrastructure/Importing/Messages/PlacesProcessingBatchMessage.cs
@@ -1,5 +1,5 @@
 ﻿using Pinventory.Pins.Infrastructure.Messages;
 
-namespace Pinventory.Pins.Application.Importing.Messages;
+namespace Pinventory.Pins.Infrastructure.Importing.Messages;
 
 public record PlacesProcessingBatchMessage(Guid ImportId, string UserId, string ArchiveJobId, Guid[] PlaceIds) : IUserMessage;
