+++ src/Pinventory.Pins.Infrastructure/Importing/Middlewares/CurrentImportLoaderMiddleware.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 using Pinventory.Pins.Domain.Importing;
 using Pinventory.Pins.Infrastructure.Messages;
 
-namespace Pinventory.Pins.Infrastructure.Middlewares;
+namespace Pinventory.Pins.Infrastructure.Importing.Middlewares;
 
 public class CurrentImportLoaderMiddleware
 {
@@ -13,12 +14,17 @@ public class CurrentImportLoaderMiddleware
         ILogger<CurrentImportLoaderMiddleware> logger,
         CancellationToken cancellationToken = default)
     {
-        if (await dbContext.GetCurrentImportAsync(message.UserId, cancellationToken) is not { } import)
+        if (await GetCurrentImport(message.UserId) is not { } import)
         {
             logger.LogWarning("Running import not found for {UserId}", message.UserId);
             return null;
         }
 
         return import;
+
+        async Task<Import?> GetCurrentImport(string userId) =>
+            await dbContext.Imports
+                .Include(x => x.StarredPlaces)
+                .SingleOrDefaultAsync(x => x.UserId == userId && x.State == ImportState.InProgress, cancellationToken);
     }
 }
