+++ src/Pinventory.Pins.Infrastructure/Importing/Sagas/Messages/BatchCompletedMessage.cs
@@ -1,3 +1,3 @@
-﻿namespace Pinventory.Pins.Infrastructure.Sagas.Messages;
+﻿namespace Pinventory.Pins.Infrastructure.Importing.Sagas.Messages;
 
 public record BatchCompletedMessage(Guid Id, string UserId, string ArchiveJobId);
