+++ src/Pinventory.Pins.Infrastructure/Importing/Sagas/Messages/ExpectedBatchesMessage.cs
@@ -1,3 +1,3 @@
-﻿namespace Pinventory.Pins.Infrastructure.Sagas.Messages;
+﻿namespace Pinventory.Pins.Infrastructure.Importing.Sagas.Messages;
 
 public record ExpectedBatchesMessage(Guid Id, string UserId, string ArchiveJobId, int BatchesCount);
