+++ src/Pinventory.Pins.Infrastructure/Importing/Sagas/Messages/ImportProcessCompleted.cs
@@ -1,5 +1,5 @@
 ﻿using Pinventory.Pins.Infrastructure.Messages;
 
-namespace Pinventory.Pins.Infrastructure.Sagas.Messages;
+namespace Pinventory.Pins.Infrastructure.Importing.Sagas.Messages;
 
 public record ImportProcessCompleted(Guid ImportId, string ArchiveJobId, string UserId) : IUserMessage;
