+++ src/Pinventory.Pins.Infrastructure/Importing/Sagas/Messages/ImportProcessTimeout.cs
@@ -4,6 +4,6 @@ using Pinventory.Pins.Infrastructure.Messages;
 
 using Wolverine;
 
-namespace Pinventory.Pins.Infrastructure.Sagas.Messages;
+namespace Pinventory.Pins.Infrastructure.Importing.Sagas.Messages;
 
 public record ImportProcessTimeout(Guid ImportId, string UserId) : TimeoutMessage(30.Minutes()), IUserMessage;
