+++ tests/Pinventory.Pins.Import.Worker.UnitTests/DataPortability/Archive/GoogleArchiveProcessorTests.cs
@@ -8,7 +8,7 @@ using Moq;
 
 using Pinventory.Pins.Import.Worker.DataPortability.Archive;
 using Pinventory.Pins.Import.Worker.DataPortability.Archive.Dtos;
-using Pinventory.Pins.Infrastructure.Services.Downloading;
+using Pinventory.Pins.Infrastructure.Downloading;
 
 using Shouldly;
